Distinguish "Editor Built In" sessions from others and sanitize imported content.
Previous commit would only accept session data if it matched the file name. But that would mean losing old session data from before that change. So this does things differently. It instead distinguishes whole-editor saves ("Editor Built In") from content-only saves. Whole-editor sessions only import old session data if the key matches the filename, but content-only sessions can fall back to the old "lastSession" key in order to continue where we left off more easily.

In order to reduce the likelihood of accidentally saving over a whole-editor file with a content-only file, whole-editor sessions save whole-editor files on ctrl+s instead of content files.

Opened content is now sanitized. If it contains nodes at the top level that shouldn't be in a content file, it's considered "somewhat bad", changing the title to say it's "imported content" and requiring the next save to specify the file rather than just saving over the last selected file that was imported. This reduces the likelihood of unintentionally changing the content of a file when opening it, making no changes, and then saving. If the imported content contains a <div> with the "userDoc" id, that's clearly importing a whole-editor file. It'll pull out the contents of that <div> and import them correctly, but it will become an "Editor Built In" session (to reduce risk of overwriting whole-editor file with content-file on next save) as well as the above-mentioned change in title and requirement to re-specify the file on save.

Editing and saving this content as a whole-editor file is fine, and there will be nothing wrong with its content. Also, opening a correct content-only file removes the warnings from the title and reverts to having ctrl+s as a content-only save shortcut, so even with all these considerations with imports and exports the editor doesn't get into a "bad state".