First pass at auto-scrolling through changes
Feels overzealous. There's good stuff there, like maintaining selection position in viewport when changing showing level. But also, making the selected line stay in position in the viewport when moving up or down feels less clear than keeping everything else still and just moving the line.

Docs waits until you try to go past the edge of the viewport and then smoothly scrolls a few lines so that it will take a few lines to trigger the scroll again. I think that's probably a good example to follow for the next one.