Undo/Redo added; Save As and Export changed
Undo/Redo:
- Ctrl+z to Undo;
- Ctrl+shift+z to Redo;

Previously there was no undo/redo. This is functional but not perfect:
- Selection isn't restored correctly after Redo;
- Some unnecessary things included in Undo states, such as word count for each node;
- Showing level isn't changed with undo/redo, but the classes will be added to elements, hiding/unhiding them anyway;
- Could be much more compact by making special cases for promoting and demoting.

Save/Export:
- Ctrl+s to Save (in current format);
- Ctrl+shift+s to Save As (in current format);
- Alt+s to Export (to other format).

Previously, there was no way to return to a "content only" format once a file becomes "editor included". Now, Save and Save As stick with the current format, but Export (Alt+s) saves a file in the other format. Once saved in the other format, it can be opened to continue in that format.

Additionally:
- Wider threshold and margin for keeping content in view when selection gets near the edge of the viewport.