Improved undo/redo, especially selection.
Selects the changed text when undoing / redoing and scrolls to make the selection visible. Backspaces are grouped together to restore one word at a time instead of one character at a time (where possible).

Also improved backspace behaviour when an element has multiple children in it, so that it doesn't think it's at the beginning of the element when it isn't.

Should better-communicate opening an editor-built-in file so as to continue editing and saving it.